Clarify wording on failure for poll-based event sources that are not stream-based
The original phrasing "When a message is not successfully processed" seems to indicate that something happens on each failed attempt to process the message, but that is not the case. Something happens after a period of time if all attempts to process have failed or no attempts have been made. In short, when the message has not been processed successfully.

Also, it is the failed message that is sent to the Dead-Letter Queue. No information about why it failed is sent.