Update info about lambda's default VPC
Referring to https://stackoverflow.com/questions/48027737/what-are-the-default-aws-lambda-vpc-settings and some resources online, I came to understand that by default, lambda is created with a VPC which is not owned by the customer account. I sometimes confuse this default lambda VPC with the default VPC provided with an AWS account. 
I believe mentioning this explicitly in this section might be helpful to others. Please feel free to suggest an alternate place for this info