Reword incorrect flag/option account-id
Apart from removing trailing whiltespaces at some places,
the purpose of this pr is to reword a flag/option in the documentation
which says following in context of invoking lambda function from s3
bucket:
```Use the `account-id` option to ensure that only resources in your account can invoke the function.```
**aws lambda cli** doesn't provide any such flag called `account-id`,
which is misleading.
I have updated the documenation with the correct flag `source-account`.