Lambda is given additional permissions than required
Lambda is given the permissions via the AWSLambdaExecute Policy which gives access to all the S3 buckets in the current AWS Account. The Lambda simply needs s3:GetObject permissions on the source S3 bucket , s3:PutObject permissions on the target S3 bucket and write permissions around the CloudWatch Logs.

What is being done in this Lambda is a common scenario and I have given this page as a reference to a lot many and they might be blindly following the documentation to give the unnecessary permissions to the Lambda. So, the changes.