Use a less abstract Lambda handler example
The first example in the documentation of how to structure a handler is (I think!) a rare use case. It's much more likely that someone writing a Lambda function will be triggering it via an AWS service than with a string or custom type.

As it stands, the documentation has the `events` package as a footnote.

By moving the `events` package into the example code, the package is much more discoverable. In your IDE, you can hit `events.` and see a list of all the available request and response types.