Update lambda-runtimes.md
This  sentence is problematic "This saves time preparing the execution environment, and it allows you to save resources such as database connections and temporary files in the execution environment to avoid creating them every time your function runs." 

The "This" is referring to an "attempt" to re-use an environment, which is different than an environment actually being reused. The "attempt" doesn't in itself save time.

The phrase "and it allows you to save resources such as database connections " is implying that the user can take certain actions because of the attempt that is made by Lambda to re-use an environment when in reality the users actions are the same regardless of whether the execution environment is reused or not. The sentence should be more clear than Lambda is doing the work. 

The phrase "creating them every time your function runs." is awkward because when an execution environment is reused the resources do not have to be recreated for that particular invocation of the function, but it should not be said that the savings applies to "every time your functions runs"  

Thank you for taking the time to review my proposed changes.