Update with-s3-example.md
In the Event JSON (Heading: Test in the console, Point: 2. iv), for the key "arn", the value "example-bucket" was written. Ideally, it should be "my-s3-bucket" because "my-s3-bucket" is the value for the bucket name.

Also, in the test event JSON template that is provided in the "Configure test event" menu on lambda function's page, test/key is written as "test%2Fkey". So, updated that as well.