:doc: wrong API type for template compatability
So if you follow along with these instructions and actually try to test the endpoint via postman or curl you will realize that you will get a server error because the lambda function is failing when trying to parse the event json data for "httpMethod". The "httpMethod" key will only be found in events generated by API gateway when it is configured as a REST API instead of an HTTP API. You can see below the differences but basically, the microservice lambda template for acting on the DynamoDB resource is set up to work out of the box for REST API gateways and not HTTP APIs. See below for comparing the body of the HTTP API vs REST API.

HTTP API Event Data Payload structure:
=============================================================
```
"requestContext": {
        "accountId": "391741500388",
        "apiId": "lh8rxrmng8",
        "domainName": "lh8rxrmng8.execute-api.us-east-1.amazonaws.com",
        "domainPrefix": "lh8rxrmng8",
        "http": {
            "method": "GET",
             .
             .
             .
```
=============================================================


API Event Data Payload structure:
=============================================================
```
"requestContext": {
        "resourceId": "q9c11x",
        "resourcePath": "/lambda-microservice2",
        "httpMethod": "GET",
         .
         .
         .
```
 =============================================================